Allow filtering by suite and test name, not just test name.
Previous to this commit, it is only possibly to filter by parts of
the test name. However, it's possible for multiple suites to be
loaded at the same time, and these may validly define tests with the
same name.

This change allows the specified filter to match against the suite
name as well as the test name, which should make it possible to
unambiguously run a single test from a set of test suites.