Make assert_raises respect subclasses
Having assert_raises not respect subclasses leads to fragile specs,
where you can not change your code to raise a subclass of an
existing exception class without breaking your specs.

assert_raises should operate similar to a rescue block.  If you do

  assert_raises StandardError do
    raise RuntimeError
  end

that should not raise an assertion error, because RuntimeError is
a subclass of StandardError and if you do:

  begin
    raise RuntimeError
  rescue StandardError
  end

the RuntimeError will be rescued.

This breaks backward compatibility, but IMO is the correct
behavior.  If you truly want to check for a specific exception
class:

  e = assert_raises StandardError do
    raise RuntimeError
  end
  assert_instance_of e, StandardError