Adds optional delegation mechanism to extend object with a mock
I've come by certain situations where I want to mock a specific method
on an object, and verify that it is called, but as far as the other
methods go, I want them to do their normal job.

Without this delegator, I would have to mock every single method that
gets called on the object the mock is mocking, which is both tedious,
error prone and leads to a path of brittle tests.

With the delegator I can mock "on top" of this object, allowing me to
mock a specific method, making my tests more intent revealing and
smaller.

This has costs, albeit small, on performance, but they won't get
triggered unless you opt in to the functionality. It might also be
argued that it has the drawback of hiding some implicit dependencies. I
believe, however, that it is a net win.

Am I missing some other way of doing this? Is this a terrible solution
for some reason I cannot see at the moment? @zenspider, what do you
think?