Add support for shared descriptions
Shared descriptions can be used inside modules that are included
in spec classes.  Usage of modules to share specs is currently
limited to sharing individual specs, shared descriptions allow
you to share spec subclasses across multiple specs.

Currently, if you want to share spec subclasses across multiple
specs, you end up having to do something ugly like:

module SharedSpecSubclass
  def self.included(mod)
    super
    pr = proc do
      describe "some description" do
        before do
          # setup
        end

        it "should do something" do end
      end
    end
    if mod.is_a?(Class)
      mod.class_eval(&pr)
    else
      mod.extend(Module.new do
        define_method(:included) do |m|
          super(m)
          m.class_eval(&pr)
        end
      end)
    end
  end

This commit reduces the above case to:

module SharedSpecSubclass
  extend Minitest::Spec::DSL
  shared_description "some description" do
    before do
      # setup
    end

    it "should do something" do end
  end
end

The example above is not an artificial example, it is taken
directly from Sequel's specs (which I'm converting from rspec to
minitest/spec).

Keeping with minitest's general approach of "It's Just Ruby",
there is no magic here.  We just save the shared_description
arguments and blocks in an array, and when the class includes
the module with the shared descriptions, we call describe at
that point with the arguments and block.

This supports not only classes including modules with shared
descriptions, but also classes including modules that include
modules with shared descriptions (which Sequel also uses).

Note that this code does not work if you do:

  class Module
    include Minitest::Spec::DSL
  end

This is because Minitest::Spec::DSL#include will never be called
in this case.  In ruby 2.0+, you can use prepend instead of
include, but I don't think there is a workaround for ruby <2.0.

I'm not tied to this particular API, it was just the simplest
thing I could think of.  I'm fine with an alternative API that
lets you share spec subclasses.