Assert match only counts as one assertions
Why:

* The method `assert_match` was counting as two assertions, for some
reason I could not figure out from the git log.

This change addresses the need by:

* Removing one of the assertions
* Still raising an asserting error with the expected message when the
matcher does not respond to the `=~` method, even though there is no
test for that, and I don't really understand why that would be needed.
A method like `assert_output` cleary states in the documentation it will
make two assertions. The `assert_match` method does not, on the
contrary, it claims to make one.