Handle empty backtrace from external filter
In the wild the backtrace_filter can be set by external filter (Rails do
it for their silencer). This can lead to the case there the external
backtrace_filer returns empty array. Other parts of minitest count on
the fact that there will still be something in the backtrace to use, and
can lead to issues such as "undefined method `split' for nil:NilClass"
when producing deprecation warnings.

Since we already had the logic to use all the backtrace when filtering
too much, I've moved this logic to be used also for external filters.