Correctly unescape newlines in diffs
@carolineschnapp noticed that, if the code under test mistakenly uses a
single-quoted string rather than a double-quoted one, the diff shown by
assert_equal is unhelpful in the presence of newlines:

  mistake = 'hello\nworld'
  assert_equal "hello\nworld", mistake

  Failure:
  Test#test_assert_equal:
  --- expected
  +++ actual
  @@ -1,2 +1,2 @@
  -"hello
  +"hello\
   world"

Whereas the expected diff would be:

  Failure:
  Test#test_assert_equal:
  --- expected
  +++ actual
  @@ -1,2 +1 @@
  -"hello
  -world"
  +"hello\\nworld"

This is happening because of indiscriminate unescaping of the character
sequence “\n” in mu_pp_for_diff. In the case where the backslash is
itself escaped, as in the string "hello\\nworld", it represents a
literal backslash and therefore should have no effect on the subsequent
“n”.

In lieu of String#undump [1,2] being universally available to solve the
more general problem, our best bet is to use a more permissive pattern
in String#gsub to only pick out escape sequences which aren’t already
part of another escape sequence, then replace those sequences if they
happen to represent a newline.

[1] https://github.com/ruby/ruby/commit/bbec11d329f5a72fe6151ec9fb0e25ff255f2eed
[2] https://github.com/tadd/string_undump