Implicitly assert nothing is raised inside block based assertions
A somewhat common mistake is to entirely skip a block based assertion
because of an improper use of `assert_raises`, e.g.

```
assert_raises TypeError do
  assert_output "yay" do
    print "nay"
    raise TypeError
  end
end
```

In this snippet `assert_output` is totally skipped but the test passes.

Of course it should be rewritten as:

```
assert_output "yay" do
  assert_raises TypeError do
    print "nay"
    raise TypeError
  end
end
```

But it's fairly easy to overlook this mistake.