Don't pass implicit blocks for method_missing stubs
Switch from `send :define_method` to `class_eval` for stubbed methods
that would otherwise by handled by `method_missing`

With the previous implementation, the block passed to `stub` is in
scope during `define_method`, which means that once a method on an
object has been stubbed, subsequent method calls will always receive
that implicit block.

I've added a new test that presents this behavior with the existing
implementation, and works properly with this change.