- Apparently adding real kwarg support to mocks/stubs broke some code. Fixed.
Short version: kludgy "last arg is a hash == kwargs" is kludgy and prevents the last arg from really being a hash... Real kwarg support is going to require you to be explicit about kwargs vs hashes.

In the meantime, an environment variable `MT_KWARGS_HACK=1` will allow the old behavior to _mostly_ behave the same. If that variable is set, AND you don't pass any use of kwargs to `expect`, AND the last arg is a Hash, then it'll use those as kwargs. If you have the ENV set and explicitly declare kwargs and then it'll warn once about having a hash as the last arg.

[git-p4: depot-paths = "//src/minitest/dev/": change = 13443]