Idea for manually registering plugins
This is just a straw man commit.  It introduces a method called
`Minitest.add_plugin`.  You pass an object to `add_plugin` and that
object is registered as a plugin.

This commit converts Minitest::PrideIO to use this interface, but that
change is only for demonstration purposes.  The existing plugin
infrastructure should be backwards compatible.

The advantages of this change are:

1. Plugins don't need to monkeypatch Minitest to register themselves
2. Plugins can manually be added (IOW we don't have to scan gems)