Use Kernel#warn uplevel argument for nicer warnings
Rather than to include the caller location in the warning message,
Minitest can pass the `uplevel` argument to `Kernel#warn` so that
the wrning is emited from the caller.

Before:

```
DEPRECATED: Use assert_nil if expecting nil from /tmp/min.rb:9. This will fail in Minitest 6.
```

After:

```
/tmp/min.rb:9: warning: DEPRECATED: Use assert_nil if expecting nil. This will fail in Minitest 6.
```

This format is understood by various editor plugins and such, allowing
a quick jump to source an similar features.